refactor(add): extract security audit rendering
