+++ src/security-audit.ts
@@ -1,0 +1,75 @@
+import pc from 'picocolors';
+import { type AuditResponse, type PartnerAudit } from './telemetry.ts';
+
+export function riskLabel(risk: string): string {
+  switch (risk) {
+    case 'critical':
+      return pc.red(pc.bold('Critical Risk'));
+    case 'high':
+      return pc.red('High Risk');
+    case 'medium':
+      return pc.yellow('Med Risk');
+    case 'low':
+      return pc.green('Low Risk');
+    case 'safe':
+      return pc.green('Safe');
+    default:
+      return pc.dim('--');
+  }
+}
+
+export function socketLabel(audit: PartnerAudit | undefined): string {
+  if (!audit) return pc.dim('--');
+  const count = audit.alerts ?? 0;
+  return count > 0 ? pc.red(`${count} alert${count !== 1 ? 's' : ''}`) : pc.green('0 alerts');
+}
+
+/** Pad a string to a given visible width (ignoring ANSI escape codes). */
+export function padEnd(str: string, width: number): string {
+  const visible = str.replace(/\x1b\[[0-9;]*m/g, '');
+  const pad = Math.max(0, width - visible.length);
+  return str + ' '.repeat(pad);
+}
+
+export function buildSecurityLines(
+  auditData: AuditResponse | null,
+  skills: Array<{ slug: string; displayName: string }>,
+  source: string
+): string[] {
+  if (!auditData) return [];
+
+  const hasAny = skills.some((s) => {
+    const data = auditData[s.slug];
+    return data && Object.keys(data).length > 0;
+  });
+  if (!hasAny) return [];
+
+  const nameWidth = Math.min(Math.max(...skills.map((s) => s.displayName.length)), 36);
+
+  const lines: string[] = [];
+  const header =
+    padEnd('', nameWidth + 2) +
+    padEnd(pc.dim('Gen'), 18) +
+    padEnd(pc.dim('Socket'), 18) +
+    pc.dim('Snyk');
+  lines.push(header);
+
+  for (const skill of skills) {
+    const data = auditData[skill.slug];
+    const name =
+      skill.displayName.length > nameWidth
+        ? skill.displayName.slice(0, nameWidth - 1) + '\u2026'
+        : skill.displayName;
+
+    const ath = data?.ath ? riskLabel(data.ath.risk) : pc.dim('--');
+    const socket = data?.socket ? socketLabel(data.socket) : pc.dim('--');
+    const snyk = data?.snyk ? riskLabel(data.snyk.risk) : pc.dim('--');
+
+    lines.push(padEnd(pc.cyan(name), nameWidth + 2) + padEnd(ath, 18) + padEnd(socket, 18) + snyk);
+  }
+
+  lines.push('');
+  lines.push(`${pc.dim('Details:')} ${pc.dim(`https://skills.sh/${source}`)}`);
+
+  return lines;
+}
